split peregrine spec
