+++ nodes/parachain/src/chain_spec/peregrine/develop.rs
@@ -1,0 +1,97 @@
+use sc_service::ChainType;
+use sp_core::sr25519;
+
+use peregrine_runtime::WASM_BINARY;
+use runtime_common::{
+	constants::{kilt_inflation_config, staking::MinCollatorStake, KILT, MAX_COLLATOR_STAKE},
+	AuthorityId,
+};
+
+use crate::chain_spec::{
+	get_account_id_from_seed, get_from_seed, get_properties, peregrine::ChainSpec, Extensions, DEFAULT_PARA_ID,
+};
+
+use super::testnet_genesis;
+
+pub fn get_chain_spec_dev() -> Result<ChainSpec, String> {
+	let properties = get_properties("PILT", 15, 38);
+	let wasm = WASM_BINARY.ok_or("No WASM")?;
+
+	Ok(ChainSpec::from_genesis(
+		"KILT Peregrine Develop",
+		"kilt_peregrine_dev",
+		ChainType::Local,
+		move || {
+			testnet_genesis(
+				wasm,
+				vec![
+					(
+						get_account_id_from_seed::<sr25519::Public>("Alice"),
+						None,
+						2 * MinCollatorStake::get(),
+					),
+					(
+						get_account_id_from_seed::<sr25519::Public>("Bob"),
+						None,
+						2 * MinCollatorStake::get(),
+					),
+				],
+				kilt_inflation_config(),
+				MAX_COLLATOR_STAKE,
+				vec![
+					(
+						get_account_id_from_seed::<sr25519::Public>("Alice"),
+						get_from_seed::<AuthorityId>("Alice"),
+					),
+					(
+						get_account_id_from_seed::<sr25519::Public>("Bob"),
+						get_from_seed::<AuthorityId>("Bob"),
+					),
+				],
+				vec![
+					(get_account_id_from_seed::<sr25519::Public>("Alice"), 10000000 * KILT),
+					(get_account_id_from_seed::<sr25519::Public>("Bob"), 10000000 * KILT),
+					(get_account_id_from_seed::<sr25519::Public>("Charlie"), 10000000 * KILT),
+					(get_account_id_from_seed::<sr25519::Public>("Dave"), 10000000 * KILT),
+					(get_account_id_from_seed::<sr25519::Public>("Eve"), 10000000 * KILT),
+					(get_account_id_from_seed::<sr25519::Public>("Ferdie"), 10000000 * KILT),
+					(
+						get_account_id_from_seed::<sr25519::Public>("Alice//stash"),
+						10000000 * KILT,
+					),
+					(
+						get_account_id_from_seed::<sr25519::Public>("Bob//stash"),
+						10000000 * KILT,
+					),
+					(
+						get_account_id_from_seed::<sr25519::Public>("Charlie//stash"),
+						10000000 * KILT,
+					),
+					(
+						get_account_id_from_seed::<sr25519::Public>("Dave//stash"),
+						10000000 * KILT,
+					),
+					(
+						get_account_id_from_seed::<sr25519::Public>("Eve//stash"),
+						10000000 * KILT,
+					),
+					(
+						get_account_id_from_seed::<sr25519::Public>("Ferdie//stash"),
+						10000000 * KILT,
+					),
+				],
+				DEFAULT_PARA_ID,
+				get_account_id_from_seed::<sr25519::Public>("Alice"),
+			)
+		},
+		vec![],
+		None,
+		None,
+		None,
+		Some(properties),
+		Extensions {
+			relay_chain: "rococo_local_testnet".into(),
+			para_id: DEFAULT_PARA_ID.into(),
+		},
+	))
+}
